Discord Bot  - Rev Learning Application
Simple Discord bot that I use in my private Discord server to learn. If you intend to use this script for your own bot, please create a database in the same folder where the script is located for it to work properly. You must also have a way to manipulate data inside the database. Create a table with the data you desire to insert, and make sure to include your Discord bot token. Do not name a column 'token'; instead, use a different name like 'bot_token' to avoid errors. Before starting the script, please change the tag in the initialization function to match your bot's tag or another identifier for your desired token.